requirements.txt
# Dependencias del Proyecto - Análisis Genómico E. coli K-12 MG1655

# Core científico
biopython==1.86
numpy==2.4.2
pandas==3.0.0

# Visualización
matplotlib==3.10.8
seaborn==0.13.2

# Utilidades
requests>=2.31.0
